+++ daikoku/javascript/src/components/utils/sidebar/index.js
@@ -1,0 +1,1 @@
+export * from './SideBar.js';
