+++ daikoku/javascript/src/components/utils/sidebar/panels/index.js
@@ -1,0 +1,4 @@
+export * from './AddPanel';
+export * from './GuestPanel';
+export * from './SearchPanel';
+export * from './SettingsPanel';
